feat: moved all actions except for long specific ones
